+++ package/com.unity.formats.usd/Runtime/Scripts/IO/Materials/StandardShaderExporter.cs
@@ -200,11 +200,7 @@ namespace Unity.Formats.USD
                 surface.metallic.defaultValue = .5f;
             }
 
-            // Gross heuristics to detect workflow.
-            if (material.IsKeywordEnabled("_SPECGLOSSMAP")
-                || material.HasProperty("_SpecColor")
-                || material.HasProperty("_SpecularColor")
-                || material.shader.name.ToLower().Contains("specular"))
+            if (ShouldUseSpecularWorkflow(material))
             {
                 if (material.HasProperty("_SpecGlossMap") && material.GetTexture("_SpecGlossMap") != null)
                 {
@@ -418,8 +414,11 @@ namespace Unity.Formats.USD
 
             if (mat.HasProperty("_OcclusionMap") && mat.GetTexture("_OcclusionMap") != null)
             {
+                // Occlusion maps in built-in standard shaders have the same data in every channel.
+                // In URP, the occlusion is packed into the G channel of a mask map.
+                // https://docs.unity3d.com/Packages/com.unity.render-pipelines.universal@16.0/manual/lit-shader.html#channel-packing
                 var newTex = SetupTexture(scene, usdShaderPath, mat, surface, Vector4.one, destTexturePath,
-                    "_OcclusionMap", "r");
+                    "_OcclusionMap", "g");
                 surface.occlusion.SetConnectedPath(newTex);
             }
 
@@ -458,6 +457,20 @@ namespace Unity.Formats.USD
             Cutout = 1,
             Fade = 2, // Old school alpha-blending mode, fresnel does not affect amount of transparency
             Transparent = 3 // Physically plausible transparency mode, implemented as alpha pre-multiply
+        }
+
+        // return true if the material should use specular workflow
+        private static bool ShouldUseSpecularWorkflow(Material material)
+        {
+            // Workflow mode of 0 is Specular and 1 is Metallic
+            if (material.HasProperty("_WorkflowMode"))
+                return (material.GetFloat("_WorkflowMode") == 0.0f);
+
+            // Gross heuristics to detect specular workflow if workflow mode not set
+            return material.IsKeywordEnabled("_SPECGLOSSMAP")
+                || material.HasProperty("_SpecColor")
+                || material.HasProperty("_SpecularColor")
+                || material.shader.name.ToLower().Contains("specular");
         }
     }
 }