+++ package/com.unity.formats.usd/CHANGELOG.md
@@ -2,9 +2,11 @@
 
 ## [3.0.0-exp.5] - Unreleased
 ### Features
+- Added basic URP Import support.
 
 ### Bug Fixes
 - Fixed export for Prefabs from the Project Window.
+- Fixed export of URP base and occlusion maps for Lit shader.
 
 ### Changed
 - Renamed USD.md in the Documentation folder to index.md to match documentation generation standards for landing pages.